rte: add PSU commands and documentation
Somehow fixes: https://github.com/Dasharo/osfv-scripts/issues/69

I think in the end we do not need that much logic in this CLI tool, but
the PSU control was definitely missing. We provide a single interface
to control the PSU, no matter what power control (relay, Sonoff) is in
place.

It has been added and exposed to the OSFV via RF libs. The CLI tool
should not know much about the firmware behavior, such as whether
the DUT will power on when PSU is enabled, or not. This is often
configurable, even at runtime, and should be handled in test environment
instead. In many cases, the CLI tool may even operate with no firmware
at all, or with bricked DUT.

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>